requirements.txt
#pkg-resources==0.0.0
simplejson
regex
chardet
dill
git-python